Assign perfd to main class and start rmt_storage on boot
perfd is not assigned to any service class. This is preventing FDE
to unmount userdata partition when encryption is started. Assigning
main class to perfd so that it can release all of its resources once
FDE wants to unmount userdata.

Starting rmt_storage service on encrypted device on boot as if start
of  rmt_storage is delated then modem may be timed out resulting in
watchdog bite.

CRs-Fixed: 787884

Change-Id: I7a5534f9a9417f2d44333122abc2abf9d83ad9a6